aux window - skip hidden windows in timeout scheduling
